requirements.txt
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pymongo==4.10.1
python-dotenv==1.0.1
typing_extensions==4.12.2
Werkzeug==3.1.3
yfinance==0.2.36
unittest2==1.1.0
